+++ partials/headers.php
@@ -57,7 +57,7 @@
 				<a class="menu_item" href="index.php">Home</a> 
 				<a class="menu_item" href="#">Shop</a>
 				<a class="menu_item" href="about.php">About</a>
-				<a class="menu_item" href="#">Terms</a>
+				<a class="menu_item" href="terms.php">Terms</a>
 				<a class="menu_item" href="contact.php">Contact</a>
 				<div class="flex-panel"></div>
 			</div>