+++ terms.php
@@ -1,0 +1,148 @@
+<!DOCTYPE html>
+<html>
+<head>
+
+	<meta charset="UTF-8">
+	<meta name="description" content="Sweet Elite is a dating sims inspired by the popular Japanese Otome.Flirt with students and uncover Arlington Academy's dark secrets.Choose your own story!">
+	
+	<title>Sweet Elite: Flirt and Uncover the Secrets of Arlington Academy!</title>
+
+	<link href="css/se-stylesheet.css" rel="stylesheet" type="text/css">
+	<link href='https://fonts.googleapis.com/css?family=Roboto:400,100,100italic,300,300italic,400italic,500,500italic,700,700italic,900,900italic' rel='stylesheet' type='text/css'>
+	<link href="https://fonts.googleapis.com/css?family=Great+Vibes" rel="stylesheet" type='text/css'>
+
+
+	<link rel="icon" type="image/ico" href="images/favicon.png"/>
+	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+
+</head>
+
+<body>
+
+	<?php include_once 'partials/headers.php' ?>
+
+	<div class = "main_content">
+        <div class = "card-nomargin add_padding">
+            <h1>Terms and Conditions</h1>
+
+            <h2>GENERAL INTRODUCTION</h2>
+            <p><b>Sweet Elite</b> is a virtual mystery game and dating simulation. You will incarnate the character of a student on a scholarship to a very prestigious high school. There are many story and dialog options offered to you, each having different impacts on the plots and on your relationships with the characters.</p></br>
+            <p>The goal of Sweet Elite is to solve the story’s main mystery while obtaining a date with your favorite character and end-of-chapter illustrations.</p></br>
+            <p>To achieve the game’s goal, the player must complete different objectives, thus enabling their character to advance throughout the story and build relationships with the characters. The player will also have the opportunity to customize their character with many virtual items of clothing</p></br>
+            <p>These TERMS AND CONDITIONS outline the different uses of the website http://sweetelitegame.com as well as all of its subdomains. They govern the main terms and conditions that the VISITOR must adhere to and accept in order to register and play the game.</p></br>
+            <p>The TERMS AND CONDITIONS are a binding agreement between the following parties:</p></br>
+            <p>DULCET GAMES INC., a Federal Corporation registered under the Canada Business Corporations Act - 2016-12-01, Corporation Number: 995484-8</p></br>
+            <p>And</p></br>
+            <p>You, the USER of the website http://sweetelitegame.com</p></br>
+            <p>The TERMS AND CONDITIONS <b>MUST BE READ AND ACCEPTED PRIOR TO REGISTRATION ON THE PLATFORM.</b> Any usage of the website sweetelitegame.com and all of its subdomains by you will be assumed to constitute an unconditional acceptance of the TERMS AND CONDITIONS. <b>If you do not agree with the following, please leave and stop use of the PLATFORM immediately.</b></p></br>
+
+            <h2>ARTICLE 1 - DEFINITIONS</h2>
+            <p>PLATFORM: Refers to the internet website accessible at http://sweetelitegame.com and all of its subdomain and pages.</p></br>
+            <p>USER: Refers to all MEMBERS and VISITORS and includes you.</p></br>
+            <p>VISITOR: Any person who accesses and navigates on the PLATFORM. May or may not be a MEMBER.</p></br>
+            <p>MEMBER: Refers to any USER that has registered on the PLATFORM to gain access to the services and products offered by DULCET GAMES INC.</p></br>
+            <p>ACCOUNT: Refers to the data supplied to the MEMBER in the form of an online interface. This interface will be generated based on the data given to DULCET GAMES INC. from the USER.</p></br>
+
+            <h2>ARTICLE 2 - ELIGIBLE USERS</h2>
+            <p>Registration to the PLATFORM is free. Individuals can open one ACCOUNT per email address. For full enjoyment of the PLATFORM and its services, DULCET GAMES INC. recommends each individual to open only one unique ACCOUNT.</p></br>
+            <p>Registration to the PLATFORM is open to any individuals of the age of 18 years old or above.</p></br>
+            <p>Sweet Elite treats subject matters that <b>may not be suitable for individuals under the age of 14.</b>If you are a parent or legal guardian, please read this page for further information on the game before giving authorization to your child to play.</p></br>
+            <p>All minors (defined as any individual under the age of 18) must obtain authorization from their parent or legal guardian before registration. <b>You are prohibited from using this website if you are a minor without authorization from your parent or legal guardian.</b></p></br>
+            <p>The purchase of “credits” or any other paid service or product is restricted to individuals over the age of 18 and to minors that have received authorization from their parent or legal guardian.</p></br>
+            <p>DULCET GAMES INC. reserves the right to refuse registration to any individual for failure to adhere to the TERMS AND CONDITIONS or if the individual’s intents or background may cause harm to the company. This can be done with no liability whatsoever.</p></br>
+            <p>The MEMBER is responsible for keeping their account information (email address and password) safe. Any use of the ACCOUNT by someone else but the associated MEMBER will not be deemed the responsibility of DULCET GAMES INC. </p></br>
+            <p>We recommend that MEMBERS use “screen names” as usernames in order to protect their privacy.</p></br>
+
+            <h2>ARTICLE 3 - CREDITS PURCHASED</h2>
+            <p>Purchased credits are completed financial transactions valid for use solely on the PLATFORM. Once purchased they are NON-refundable and have NO cash redemption value.</p></br>
+            <p>The credits and ACCOUNTS are non-transferable.</p></br>
+            <p>If a MEMBER chooses to close their ACCOUNT, no matter what the cause or reason, the credits will be lost and non-refundable.</p></br>
+
+            <h2>ARTICLE 4 - PROHIBITIONS</h2>
+            <p>The following is a list of activities or content that is prohibited at all times and on all subdomains of the website www.sweetelitegame.com. DULCET GAMES INC. reserves the right to take legal action against anyone who violates this provision and/or remove them, without any limitation, from the PLATFORM.</p></br>
+                <li>Promotion of violence: any MEMBER who is libelous, defamatory, obscene, racist, or hateful; defends war crimes or pedophiles, incites murder or suicide, incites discrimination or hatred, promotes racism, bigotry, hatred or physical harm of any kind against any group or individual, reproduces (or promotes the reproduction of) a copyrighted work without authorization, or is otherwise contrary to any applicable law, rule or regulation.</li>
+                <li>Harassment or causing harassment to another individual</li>
+                <li>Solicitation of personal information from any individual under the age of 18</li>
+                <li>Promotion of false or misleading information about DULCET GAMES INC.</li>
+                <li>Promotion of illegal activities</li>
+                <li>Promotion of chain letters or spam</li>
+                <li>Participation in illegal activities</li>
+                <li>Attempting to impersonate another MEMBER or individual</li>
+                <li>Covering or modifying the advertisement banners on your profile page or any page using HTML/CSS or other means.</li>
+                <li>Using the ACCOUNT, username, password or any other information of another MEMBER</li>
+                <li>Using the PLATFORM for any purpose other than your own personal use</li>
+                <li>Participation in commercial activities (including advertisements and sales) on the PLATFORM without the written consent of DULCET GAMES INC.</li>
+                <li>Divulgation of personal information without that individual’s consent</li>
+                <li>Promotion of nudity, violence or subject matter that would be present on an adult website</li>
+            </br><p>If you are interested in using Sweet Elite for organizational or commercial purposes, please contact contact@sweetelitegame.com</p></br>
+
+            <h2>ARTICLE 5 - INTELLECTUAL PROPERTY OF DULCET GAMES INC.</h2>
+            <p>The trademarks, logos, slogans, graphics, photographs, animations, videos, software and text contained on the PLATFORM are the exclusive property of DULCET GAMES INC. and may NOT be reproduced, used or represented without the express permission of DULCET GAMES INC., under penalty of law.</p></br>
+            <p>Any total or partial reproduction, use or representation of the PLATFORM, by any means whatsoever, without the express permission of DULCET GAMES INC. is prohibited and shall constitute an infringement punishable.</p></br>
+            <p>IT IS STRICTLY PROHIBITED TO EXTRACT CONTENT OF DULCET GAMES INC.’s DATABASE AND/OR MAKE IT AVAILABLE TO THE PUBLIC.</p></br>
+
+            <h2>ARTICLE 6 - LIABILITY OF THE USER</h2>
+            <p>Any use of the PLATFORM and its related websites and services is agreed by you to be done at your own risk. By being a USER, you assume ALL risk associated with the client-based performance and use of the PLATFORM. DULCET GAMES INC. is not responsible for damage to your equipment, loss of data or unavailability/interruption of operations. DULCET GAMES INC. does not guarantee any specific outcomes from the use of the PLATFORM.</p></br>
+            <p>DULCET GAMES is not liable for the following:</p></br>
+                <li>Temporary unavailability/lack of access to the PLATFORM due to technical maintenance</li>
+                <li>Virus attacks</li>
+                <li>Abnormal or illegal use of the website by a USER. The USER is solely responsible for damage to third parties.</li>
+            <p>You agree to indemnify and hold DULCET GAMES INC. and their respective shareholders, directors, officers, agents, partners, and employees, harmless from any loss, liability, claim, or demand, including reasonable attorneys' fees, expenses and costs, made by any third party related to or arising out of your use of the PLATFORM in violation of this Agreement, arising from a breach of this Agreement, and/or if any Content that you post on the PLATFORM or through the PLATFORM causes us or any such other indemnified person to be liable to another.</p></br>
+            <p>If the situation arises, it is up to the USER to prove the fault of DULCET GAMES INC. in order to challenge its liability, which will be strictly limited to direct damages. IN NO EVENT SHALL DULCET GAMES INC. BE LIABLE TO YOU OR ANY THIRD PARTY FOR ANY INDIRECT, CONSEQUENTIAL, EXEMPLARY, INCIDENTAL, SPECIAL OR PUNITIVE DAMAGES, INCLUDING LOST PROFIT DAMAGES ARISING FROM YOUR USE OF THE PLATFORM.</p></br>
+
+            <h2>ARTICLE 7 - OLIGATIONS OF THE USER</h2>
+            <p>When using the PLATFORM, the user agrees to maintain public order and comply to all of the rules laid out in the TERMS AND CONDITIONS. The user must:</p></br>
+                <li>Be honest and provide accurate information to DULCET GAMES INC. and other users</li>
+                <li>Respect the rights of DULCET GAMES INC. and other users</li>
+                <li>Use the PLATFORM in accordance with the TERMS AND CONDITIONS</li>
+                <li>Respect the privacy of other users</li>
+                <li>Not refer users towards competing sites and games</li></br>
+
+            <h2>ARTICLE 8 - OLIGATIONS OF DULCET GAMES INC.</h2> 
+            <p>DULCET GAMES INC. is committed to ensuring constant access to the PLATFORM for the users, 7 days a week, 24 hours a day.</p></br>
+            <p>DULCET GAMES INC. will remove all illegal content from the PLATFORM as soon as it has been reported.</p></br>
+            <p>DULCET GAMES INC. has a general obligation of providing the services promised. The company has no other signification obligation.</p></br>
+
+            <h2>ARTICLE 9 - TERMINATION & MODIFICATION OF TERMS</h2>
+            <p>Each MEMBER may terminate its ACCOUNT by requesting a closure to DULCET GAMES INC. This request shall be deemed effective the business day following the receipt of the request by DULCET GAMES INC. Upon termination, all data concerning the USER will be removed.</p></br>
+            <p>DULCET GAMES INC. may terminate a USER’s ACCOUT without notice or demand if:</p></br>
+                <li>The USER fails to comply with these TERMS AND CONDITONS</li>
+                <li>The ACCOUNT remains inactive for more than 3 years continuously.</li></br>
+            <p>Any Termination executed by DULCET GAMES INC. carries no legal liability whatsoever.</p></br>
+            <p>DULCET GAMES INC. reserves the right to modify these TERMS AND CONDITIONS at any time. Changes will take effect immediately upon publication on the PLATFORM. Any registration after the date of publication of modified TERMS AND CONDITIONS will be deemed as having been made after the consenting to the new version.</p></br>
+            <p>Any MEMBER registered before the modifications reserves the right to terminate their ACCOUNT at any time.</p></br>
+
+            <h2>ARTICLE 10 - GENERAL PRIVACY PROVISION</h2>
+            <p>Privacy is important to DULCET GAMES INC. This Provision applied to all VISITORS of www.sweetelitegame.com. Our Privacy Policy is designed to assist you in understanding how we collect and use personally identifiable information (“Personal Information”) that you provide to us and to assist you in making informed decisions when using our Website. We will not use or share your Personal Information with anyone except as described in this Privacy Policy.</p></br>
+            <p><b>As a general rule, you should avoid disclosing any personal information that you do not wish to make public to other USERS.</b></p></br>
+            <p>In order to make use of the services offered by the PLATFORM, we may ask you for Personal Information. Personal Information includes, for example, your email address, which you are asked to provide when registering with DULCET GAMES INC.</p></br>
+            <p>When you visit the PLATFORM we may collect information regarding your visit from your computer. Such information helps us to determine information about how you found our Website, the date that you visited our Website and search engines that may have led you to our Website. We collect this information in order to continually improve and enhance the functionality of www.sweetelitegame.com</p></br>
+            <p>If you email us for support, we may ask you for information such as your name and email address so we can respond to your questions and comments. If you choose to correspond with us via email, we may retain the content of your email messages, your email address and our response to you. In certain cases, and with your permission, we may post content from your emails to us on the BEEMOOV Website.</p></br>
+            <p>Your Personal Information may be retained for the period of your ongoing use plus up to three years after you expressly terminate your membership or your account is dormant for a period of six months. This website is not directed to children under the age of 14 years. If you are under the age of 14, you are not allowed to use this site. We recommend that, if you are under 14, you should be smart and not provide any information on your identity, physical location or online contact information unless you do so under your parent’s consent and supervision.</p></br>
+            <p>DULCET GAMES INC. may use “cookies” on or in connection with our Website to maintain information about you. A cookie is a very small text document, which often includes an anonymous unique identifier. DULCET GAMES INC. employs the use of both “session” cookies and “persistent” cookies. A persistent cookie remains on your hard drive after you close your browser. Persistent cookies may be used by your browser on subsequent visits to the Website. Persistent cookies can be removed by following your web browser’s directions. A session cookie is temporary and disappears after you close your browser.</p></br>
+            <p>When you visit www.sweetelitegame.com, our computer may ask your computer for permission to employ the use of a session or persistent cookie. Our Website will then send a cookie to your browser if your browser’s preferences allow it, but (to protect your privacy) your browser only permits our Website to access the cookies it has already sent to you and not the cookies sent to you by other websites.</p></br>
+            <p>Most web browsers can be adjusted to inform you when a cookie has been sent to you and provide you with the opportunity to refuse that cookie. However, refusing a cookie may, in some cases, preclude you from using, or negatively impact the display or function of, our Website or certain areas or features of our Website.</p></br>
+            <p>DULCET GAMES INC. may use third-party partners to help us operate and maintain our Website and deliver our products and services. We may also share your Personal Information and Website Usage Information with our service providers and other third parties (“Affiliated Parties”) that provide products or services for or through this Website or for our business (such as website or database hosting companies, address list hosting companies, communications providers, email service providers, analytics companies, distribution companies, fulfillment companies, credit card processing companies and other similar service providers that use such information on our behalf).</p></br>
+            <p>Access to our database is granted only to our technicians on a case-by-case basis to troubleshoot specific technical issues, or as may be required by law.</p></br>
+            <p>While DULCET GAMES INC. will do its best to protect the security and integrity of sensitive Personal Information provided to this Website, due to the inherent nature of the Internet as an open global communications vehicle, we cannot guarantee that information, during transmission through the Internet or while stored on our systems or otherwise in our care, will be safe from intrusion by others, such as hackers.</p></br>
+            <p>As we update and expand our Website, this Privacy Policy may change, so please check this web page for updates to our Privacy Policy.</p></br>
+
+            <h2>ARTICLE 11 - GENERAL PROVISION</h2>
+            <p>By registering, you agree that DULCET GAMES may contact you relating to your ACCOUNT or your use on the PLATFORM.</p></br>
+            <p>By being a USER, we may contact you for marketing, surveys, promotional or other reasons where you have previously consented. You may notify us to opt-out of such communications.</p></br>
+            <p>No other guidelines or documents can create obligations not included in the TERMS AND CONDITIONS unless agreed in writing by both Parties.</p></br>
+            <p>This document sets forth the entire and exclusive expression of the agreements and obligations of the Parties. No other guidelines or documents can create obligations not included in the TERMS AND CONDITIONS unless agreed in writing by both Parties. The fact that one of the Parties did not require the enforcement of any provision of these Terms, whether permanently or temporarily, will in no way be considered a waiver of such provision.</p></br>
+            <p>The titles contained herein are for convenience of reference only and shall not be used to interpret the meaning of these TERMS AND CONDITIONS. If any provision were to be deemed invalid under a statute or regulation in force and/or a court decision, it will be disregarded but will not affect the validity of the remaining provisions which shall remain fully applicable.</p></br>
+            <p>EACH OF THE PARTIES HEREBY WAIVES TRIAL BY JURY IN ANY ACTION ARISING OUT OF MATTERS RELATED TO THIS AGREEMENT. THIS WAIVER AND THE OTHER CONSENTS REGARDING RESOLUTION OF DISPUTES IS INFORMED AND VOLUNTARY.</p></br>
+            <p>Unless otherwise stated, all references to a law or regulation refers to the laws of Canada.</p></br>
+            <p>Sweet Elite is a work of fiction. Names, characters, businesses, places, events and incidents are either the products of the author’s imagination or used in a fictitious manner. Any resemblance to actual persons, living or dead, or actual events is purely coincidental, and does not place any liability whatsoever on DULCET GAMES INC.</p></br>
+            <p>© 2018 DULCET GAMES INC. All rights reserved.</p>
+		</div>
+	</div>
+
+    <?php include_once 'partials/footers.php' ?>
+
+</body>
+
+</html>
+